% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cnv}
\alias{data_cnv}
\title{Genome data.}
\format{An object of class \code{matrix} with 2997 rows and 676 columns.}
\source{
\url{https://cancergenome.nih.gov}
}
\usage{
data_cnv
}
\description{
A matrix that contains genome data.
}
\keyword{datasets}

