% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_trscr}
\alias{data_trscr}
\title{Transcriptome data.}
\format{An object of class \code{data.frame} with 2997 rows and 505 columns.}
\source{
\url{https://cancergenome.nih.gov}
}
\usage{
data_trscr
}
\description{
A matrix that contains transcriptome data.
}
\keyword{datasets}

