% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_grp}
\alias{exp_grp}
\title{Sample metadata.}
\format{An object of class \code{data.frame} with 743 rows and 13 columns.}
\source{
\url{https://cancergenome.nih.gov}
}
\usage{
exp_grp
}
\description{
A dataframe of metadata concerning samples (`data_*` columns).
}
\keyword{datasets}

