% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_list}
\alias{gene_list}
\title{Gene metadata.}
\format{An object of class \code{numeric} of length 1.}
\source{
\url{https://cancergenome.nih.gov}
}
\usage{
gene_list
}
\description{
A dataframe of metadata concerning genes (bed formatted, `data_trscr` and `data_cnv` rows).
}
\keyword{datasets}

