% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pf_meth}
\alias{pf_meth}
\title{Methylome probes metadata.}
\format{An object of class \code{numeric} of length 1.}
\source{
\url{https://cancergenome.nih.gov}
}
\usage{
pf_meth
}
\description{
A dataframe of metadata concerning methylome probes (`data_meth` rows).
}
\keyword{datasets}

