% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_meth}
\alias{data_meth}
\title{Methylome data.}
\format{An object of class \code{data.frame} with 25344 rows and 618 columns.}
\source{
\url{https://cancergenome.nih.gov}
}
\usage{
data_meth
}
\description{
A matrix that contains methylome data.
}
\keyword{datasets}

